adding last phase

+++ src/main/java/gr/patras/carnival/game/data/repositories/WeekRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 public interface WeekRepository
         extends CrudRepository<Week, Long> {
 
-    List<Week> findById(Integer id);
+    List<Week> findById(Long id);
 
 }
